Completed fakeReviewStronger
What I have done is made it so fakeReview can see if a adjective is a positive connotation or a negative connotation, It does this by seeing if the word inputted is stronger or weaker than the previous word. After this it will either output a 1 or a -1. Outputting 1 means that the adjective has a positive meaning aka "happy" and -1 means that the adjective has a negative meaning aka "sad".